initial introduced for scope
here i create a new spring_core project for explain scope when I 'scope' use an attribute on bean tag,
when scope = "prototype" then spring not create an object when .xml file loading, when call for create an obj then give a new object. when i will call. but by default scope = "singleton" this mean this object is only one object for that application.